backend/requirements.txt
# MeshSOS Backend Dependencies

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# Serial communication
pyserial>=3.5

# Meshtastic (optional, for real hardware integration)
# meshtastic includes PyPubSub as a dependency, no need to install separately
meshtastic>=2.2.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality (optional)
ruff>=0.1.0
